requirements.txt
streamlit
pandas
qrcode[pil]
pillow
phonenumbers
